pyproject.toml
[project]
name = "wikipedia-news-updater"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.119.0",
    "httpx>=0.28.1",
    "openai>=2.5.0",
    "python-dotenv>=1.1.1",
]
